requirements.txt
appnope==0.1.0
backcall==0.1.0
cycler==0.10.0
decorator==4.3.0
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
kiwisolver==1.0.1
matplotlib==3.0.1
numpy==1.15.3
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.2.0
pykalman==0.9.5
pyparsing==2.3.0
python-dateutil==2.7.5
scipy==1.1.0
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
